requirements.txt
PySide6
openai-whisper
yt-dlp
google-generativeai
torch
numpy
requests
